+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add the `Viewer.create` static method for creating a new `Viewer` instance without the `new` operator.
+- Add a new `navigation` option to show previous and next buttons (#624).
 - Add a new `autoplay` option to disable automatic image cycling when playing (#608).
 - Allow controlling transitions by action with the `transition` option (#337).
 - Add a new `preload` option to preload the next or previous image before viewing it (#313).